+++ src/index.tsx
@@ -12,6 +12,7 @@ type IncogniaType = {
   fetchInstallationId(): Promise<string>;
   setLocationEnabled(enabled: boolean): void;
   notifyAppInForeground(): void;
+  generateRequestToken(): Promise<string>;
   ConsentTypes: ConsentTypesType;
   Trial: IncogniaTrialType;
 };
@@ -103,6 +104,7 @@ export const allowConsentTypes = IncogniaModule.allowConsentTypes;
 export const denyConsentTypes = IncogniaModule.denyConsentTypes;
 export const fetchInstallationId = IncogniaModule.fetchInstallationId;
 export const setLocationEnabled = IncogniaModule.setLocationEnabled;
+export const generateRequestToken = IncogniaModule.generateRequestToken;
 export const notifyAppInForeground = () => {
   if (Platform.OS === 'android') {
     IncogniaModule.notifyAppInForeground();
@@ -143,6 +145,7 @@ export default {
   fetchInstallationId,
   setLocationEnabled,
   notifyAppInForeground,
+  generateRequestToken,
   Trial,
   ConsentTypes,
 } as IncogniaType;